added explicit json mapper test class
