+++ test/test_matador.cpp
@@ -23,6 +23,7 @@
 #include "utils/BlobTestUnit.hpp"
 #include "utils/DateTestUnit.hpp"
 #include "utils/JsonTestUnit.hpp"
+#include "utils/JsonMapperTestUnit.hpp"
 #include "utils/TimeTestUnit.hpp"
 #include "utils/FactoryTestUnit.hpp"
 #include "utils/StringTestUnit.hpp"
@@ -84,6 +85,7 @@ int main(int argc, char *argv[])
   suite.register_unit(new TimeTestUnit);
   suite.register_unit(new BlobTestUnit);
   suite.register_unit(new JsonTestUnit);
+  suite.register_unit(new JsonMapperTestUnit);
   suite.register_unit(new FactoryTestUnit);
   suite.register_unit(new StringTestUnit);
   suite.register_unit(new SequencerTestUnit);
