+++ test/utils/JsonTestUnit.hpp
@@ -10,8 +10,6 @@ public:
 
   void test_simple();
   void test_parser();
-  void test_mapper();
-  void test_failures();
 };
